feat:rename data models and change schema
